Fixed: Turning syntax highlightning on/off did not update the source code view directly.
Fixed: The indenter was not called if another indenter was selected while preview was disabled and after indenter change preview was enabled.

Changed: The parameter string needed to write the indenter config file is now only generated automaticly if the indenter is called and not for every setting change. So while preview is disabled this saves unnecessary string generation calls.